# opt edf performance