--- a/pkg/upstream/cluster/edf_test.go
+++ b/pkg/upstream/cluster/edf_test.go
@@ -1,9 +1,11 @@
 package cluster
 
 import (
+	"strconv"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"mosn.io/mosn/pkg/types"
 )
 
 func Test(t *testing.T) {
@@ -29,4 +31,82 @@ func Test(t *testing.T) {
 	ele = edfScheduler.NextAndPush(weightFunc)
 	assert.Equal(t, A, ele)
 
+}
+
+func mockHostList(count int, name string) []types.Host {
+	hosts := make([]types.Host, 0, count)
+	for i := 0; i < count; i++ {
+		hosts = append(hosts, &mockHost{
+			name: "A" + strconv.Itoa(i),
+			w:    uint32(i + 1),
+		})
+	}
+	return hosts
+}
+
+func Benchmark_edfSchduler_NextAndPush(b *testing.B) {
+	type fields struct {
+		hostCount int
+		hostName  string
+	}
+	type args struct {
+		weightFunc func(item WeightItem) float64
+	}
+	tests := []struct {
+		name   string
+		fields fields
+		args   args
+		want   interface{}
+	}{
+		{
+			name: "10-hosts",
+			fields: fields{
+				hostCount: 10,
+				hostName:  "bench-host",
+			},
+			args: args{
+				weightFunc: func(item WeightItem) float64 {
+					return float64(item.Weight())
+				},
+			},
+		},
+		{
+			name: "100-hosts",
+			fields: fields{
+				hostCount: 100,
+				hostName:  "bench-host",
+			},
+			args: args{
+				weightFunc: func(item WeightItem) float64 {
+					return float64(item.Weight())
+				},
+			},
+		},
+		{
+			name: "1000-hosts",
+			fields: fields{
+				hostCount: 1000,
+				hostName:  "bench-host",
+			},
+			args: args{
+				weightFunc: func(item WeightItem) float64 {
+					return float64(item.Weight())
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		b.Run(tt.name, func(b *testing.B) {
+			edf := newEdfScheduler(tt.fields.hostCount)
+			hosts := mockHostList(tt.fields.hostCount, tt.fields.hostName)
+			for _, h := range hosts {
+				edf.Add(h, float64(h.Weight()))
+			}
+			b.StartTimer()
+			for i := 0; i < b.N; i++ {
+				edf.NextAndPush(tt.args.weightFunc)
+			}
+			b.StopTimer()
+		})
+	}
 }