### Issues associated with this PR

#1760 

### Solutions
You should show your solutions about the issues in your PR, including the overall solutions, details and the changes. At this time, Chinese is allowed to describe these.

### UT result
Unit Test is needed if the code is changed, your unit test should cover boundary cases, corner cases, and some exceptional cases. And you need to show the UT result.

### Benchmark

```bash
$ benchstat old.txt new.txt 
name                                    old time/op    new time/op    delta
_edfSchduler_NextAndPush/10-hosts-12       214ns ±14%     157ns ± 5%  -26.35%  (p=0.000 n=10+10)
_edfSchduler_NextAndPush/100-hosts-12      323ns ± 9%     221ns ± 2%  -31.43%  (p=0.000 n=9+10)
_edfSchduler_NextAndPush/1000-hosts-12     497ns ± 5%     336ns ± 2%  -32.45%  (p=0.000 n=8+10)

name                                    old alloc/op   new alloc/op   delta
_edfSchduler_NextAndPush/10-hosts-12       0.00B          0.00B          ~     (all equal)
_edfSchduler_NextAndPush/100-hosts-12      0.00B          0.00B          ~     (all equal)
_edfSchduler_NextAndPush/1000-hosts-12     0.00B          0.00B          ~     (all equal)

name                                    old allocs/op  new allocs/op  delta
_edfSchduler_NextAndPush/10-hosts-12        0.00           0.00          ~     (all equal)
_edfSchduler_NextAndPush/100-hosts-12       0.00           0.00          ~     (all equal)
_edfSchduler_NextAndPush/1000-hosts-12      0.00           0.00          ~     (all equal)  
```
### Code Style
+ Make sure `Goimports` has run
+ Show `Golint` result